Do not reify dependencies everytime the same module is looked up
We ideally only need to reify the dependencies of a module once for the first time it is looked up. Thereafter, we can reuse the store reified values. This PR fixes that. Earlier it was reifiying the dependencies of a module and evaluating the sub tree irrespective of if the dependencies were already resolved.