Check object extensibility before adding default property.
The checks added previously were checking for `Object.isFrozen`, but
that only covers one scenario (`Object.freeze`) and it is still possible
to make an object non-extensible without freezing it.

This adds tests for frozen objects (previously passing) and a new test
for `Object.seal`'ed objects (which was failing prior to this change).